Use add_custom_command instead of execute_process for building protos.
This more closely matches how CMake is supposed to be used. It adds the
commands as build steps instead of running them as part of analyzing the
build script itself. This allows the build steps to be parallelized,
produce output in the correct place, and run only when libicing is
actually being built.

Bug: 149853706
Test: ./gradlew external:icing:assemble
Change-Id: I548cf82d5d5ed6377fd65fa8f5ed04e90bd0d4f8