Add a smoke test for androidx.
This makes sure that the distribution in the android tree keeps working.

Due to b/162892721, the libicing.so inside the aar is not used. The
libicing.so from this tree is used. So, this test serves the dual
purpose of making sure the version of icing code available here supports
the java code in androidx-appsearch.aar.

Bug: 162755601
Test: AppSearchSmokeTest
Change-Id: Ib4ed1cb9cb546146de3e85e32b10f23ca374efbc