Stop dropping nonascii/nonalnum segments.
Cherrypick a change that would remove the existing behavior of the
LanguageSegmenter to filer out non-ascii+non-alphanumeric segments. This
behavior is more appropriate for a higher level class like the
tokenizer. This will allow a consistent hanlding: punctuation and
whitespace is not indexed. All other segments are.

This is a cherrypick of cl/446831054.

Bug: 229770338
Test: builds
Change-Id: Id6056354aaea5fca1cdb0e9300065bd3f91e401b