Sync from upstream.
Descriptions:
======================================================================
Rename fullyQualifiedId to qualifiedId
======================================================================
Delete ResultState, which is dead code.
======================================================================
Implement support in the QueryVisitor for the property restrict operator (HAS)
======================================================================
Implement support in the QueryVisitor for implicit and, 'AND' and 'OR'.
======================================================================
Implement support in the QueryVisitor for NOT.
======================================================================
Implement support in the QueryVisitor for single term and single verbatim term queries.
======================================================================

Bug: 208654892
Bug: 256022027
Change-Id: I6c689dd7070ac0a9818b7425f3beab7258181049