Use the new build tools version API
Also removed the static import in the build file since it
was resolving the new method instead of the AndroidLibrary
extension methods. I think it is better than renaming the
method since this build file would be subject to similar
problems in the future if it keeps using static imports.

Bug: 277058565
Test: CI
Change-Id: Id7a41c63e521e6f1873d3042044f39e6751ab7ed