Add //apex_available:platform to libicing_tests
Soong was accidentally allowing a corner case where modules that
were not supposed to be available to the platform (due to
apex_available being set to an apex and not including
"//apex_available:platform") were being made available to the
platform as long as no apex modules depended on them.  Add
"//apex_available:platform" explicitly to libicing_tests.

Bug: 394664568
Test: compare out/soong/installs-${TARGET_PRODUCT}.mk
Change-Id: I12c81f710b9416b78b948466fcf7f5d0d7f5a77e