Icing - GetTokens re-use buffer.
PlainTokenizer alone is responsible for 10% of
allocations from Hawkeye test, see b/404607371#comment5.

Ideally, we could switch token iterator back to
"GetToken" instead of "GetTokens" and use string
views for these (the docs are out of date, and
it seemed to have an API that encouraged less
allocations before, but let's see how it's going.

This CL is written by Gemini.

From local measurements, this does appear to reduce
the number of reclaims by a few %, but this is
inconclusive evidence.

Bug: 404607371
Test: N/A
Flag: EXEMPT refactor
Change-Id: I0353dc1c3f4f2697abc66216eecc333c3ba2aa47